+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `CodeGenerator`: reorganised array validations (particularly nested arrays)
+
 ## [0.29] - 2021-03-31
 ### Changed
 - `CodeGenerator`, `Validation`, templates: optimised decimal comparisons