+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - `CodeGenerator`: reorganised array validations (particularly nested arrays)
+- templates: improve pattern validation
 
 ## [0.29] - 2021-03-31
 ### Changed