+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - templates: improve pattern validation
 - `CodeGenerator`, `Validation`, templates: optimised array length validations
 - templates: optimised generation for multiple validations on same property
+### Added
+- integration test (initial simple test of generated code)
 
 ## [0.29] - 2021-03-31
 ### Changed